+++ src/stormwind-library.lua
@@ -1,8 +1,10 @@
--- Library version = '1.1.0'
+-- Library version = '1.2.0'
 
 -- import src/Support/Arr.lua
 -- import src/Support/Bool.lua
 -- import src/Support/Str.lua
+
+-- import src/Core/Environment.lua
 
 -- import src/Core/AddonProperties.lua
 -- import src/Core/Factory.lua
@@ -17,8 +19,18 @@
 -- import src/Facades/EventHandlers/PlayerLoginEventHandler.lua
 -- import src/Facades/EventHandlers/TargetEventHandler.lua
 -- import src/Facades/Target.lua
+-- import src/Facades/Tooltips/AbstractTooltip.lua
+-- import src/Facades/Tooltips/ClassicTooltip.lua
+-- import src/Facades/Tooltips/RetailTooltip.lua
 
+-- @TODO: Move this to AbstractTooltip.lua once the library initialization callbacks are implemented <2024.05.04>
+self.tooltip = self:new('Tooltip')
+self.tooltip:registerTooltipHandlers()
+
+-- import src/Models/Item.lua
 -- import src/Models/Macro.lua
 -- import src/Models/RaidMarker.lua
+-- import src/Models/Realm.lua
+-- import src/Models/Player.lua
 
 -- import src/Views/Windows/Window.lua